method copyAndCreate was refactored to If-else block

+++ src/Task3/Level1/PathsTo.java
@@ -9,42 +9,47 @@ import java.nio.file.Paths;
 public class PathsTo {
 
 
-
     public static Path source = Paths.get ("C:\\Users\\Lenovo X220 Tablet\\IdeaProjects\\Go2IT.dat.txt"); // manualy created file in folder
     public static Path destination = Paths.get ("C:\\Users\\Lenovo X220 Tablet\\Desktop\\Go2ITnew.dat.txt");
     public static Path destination2 = Paths.get ("C:\\Users\\Lenovo X220 Tablet\\Desktop\\Go2ITnew2.dat.txt");
 
     public void existanceOfFileSourse() {  // checking if file exist at source folder
         if (Files.exists (source)) {
-            System.out.println ("The file Exist" );
-        }else {
-            System.out.println ("The file does not exist" );
+            System.out.println ("The file Exist");
+        } else {
+            System.out.println ("The file does not exist");
         }
     }
+
     public void existanceOfFileDestination() {  // checking if file exist at destination folder
         if (Files.exists (destination)) {
-            System.out.println ("The file Exist" );
-        }else {
-            System.out.println ("The file does not exist" );
+            System.out.println ("The file Exist");
+        } else {
+            System.out.println ("The file does not exist");
         }
     }
+
     public void existanceOfFileDestination2() {  // checking if file exist at second destination  folder
         if (Files.exists (destination2)) {
-            System.out.println ("The file Exist" );
-        }else {
-            System.out.println ("The file does not exist" );
+            System.out.println ("The file Exist");
+        } else {
+            System.out.println ("The file does not exist");
         }
     }
-    public void copyAndCreate () throws IOException {
-        try {
-            Files.copy (PathsTo.source, PathsTo.destination); //copy file from 'source' to 'destination' simultaneously creating file if not exist.
-//todo: refactoring of method to if else block
-        } catch (FileAlreadyExistsException e) { // in case file already exist will delete it an recreate it again
-            System.out.println ("   File 'Go2ITnew.dat' already exist(Deleting....)  ");
-            Files.delete (PathsTo.destination);  // delete file (if exist Exception : 'FileAlreadyExistsException')
-            Files.copy (PathsTo.source, PathsTo.destination);  // copy file again
+
+    public void copyAndCreate()  {
+
+        if (!Files.exists (PathsTo.destination)) {
+            System.out.println ("4. Copying and creating the file 'Go2ITnew.dat.txt'");
+            try {
+                Files.copy (PathsTo.source, PathsTo.destination);//copy file from 'source' to 'destination' simultaneously creating file if not exist.
+            } catch (IOException e) {
+                System.out.println ("The file 'Go2ITnew.dat.txt' already Exist");
+            }
         }
     }
+
+
 
     public void copyFileByStream ()throws IOException {
         InputStream in = null;
